Add files via upload
we looked at clustering, which is an unsupervised learning approach. We
use unsupervised learning to explore data, rather than for classification and prediction
purposes. In the experiment here, we didn't have topics for the news items we found on
reddit, so we were unable to perform classification. We used k-means clustering to group
together these news stories to find common topics and trends in the data.
In pulling data from reddit, we had to extract data from arbitrary websites. This was
performed by looking for large text segments, rather than a full-blown machine learning
approach. There are some interesting approaches to machine learning for this task that may
improve upon these results. In the Appendix of this book, I've listed, for each chapter,
avenues for going beyond the scope of the chapter and improving upon the results. This
includes references to other sources of information and more difficult applications of the
approaches in each chapter.
We also looked at a straightforward ensemble algorithm, EAC. An ensemble is often a good
way to deal with variance in the results, especially if you don't know how to choose good
parameters (which is always difficult with clustering).
Finally, we introduced online learning. This is a gateway to larger learning exercises,
including big data, which will be discussed in the final two chapters of this book. These
final experiments are quite large and require management of data as well as learning a
model from them.
As an extension on the work in this chapter, try implementing EAC to be an online learning
algorithm. This is not a trivial task and will involve some thought on what should happen
when the algorithm is updated. Another extension is to collect more data from more data
sources (such as other subreddits or directly from news websites or blogs) and look for
general trends.
In the next chapter, we'll step away from unsupervised learning and go back to
classification. We will look at deep learning, which is a classification method built on
complex neural networks.